src/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
easy-pil==0.1.4
frozenlist==1.2.0
idna==3.3
lxml==4.7.1
multidict==5.2.0
Pillow==8.4.0
python-docx==0.8.11
requests==2.27.1
typing_extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
